refactor: InputFormator의 기능을 각 도메인 역할에 맞게 분배후 삭제
- 음수인지 검사 -> Position로 이동
- 숫자인지 검사, 양수인지 검사 -> TryCount로 이동
